Add files via upload
Assignment 4 : Battleship game simulator. The .txt input files represent boards for each player, .in files represent moves of players and "optional" files describe the locations of long ships because of possible confusion. 
The code produces an output file which contains every round of the game and the winner.
Open the .pdf document for more details.